+++ src/hazardous/dsa/ml_dsa/mldsa65.rs
@@ -43,6 +43,7 @@
 //!
 //! # Errors:
 //! An error will be returned if:
+//! - Verification of a signature, message and context failed.
 //! - [`getrandom::fill()`] fails during signing.
 //! - [`getrandom::fill()`] fails during [`KeyPair::generate()`].
 //! - `mu` is not `64` bytes.
@@ -314,7 +315,7 @@ impl VerifyingKey {
 /// ML-DSA-65 keypair.
 pub struct KeyPair {
     seed: Seed,
-    signing_key: SigningKey,
+    pub(crate) signing_key: SigningKey,
     pub(crate) verifying_key: VerifyingKey,
 }