+++ src/high_level/signer.rs
@@ -1,0 +1,127 @@
+// MIT License
+
+// Copyright (c) 2026 The orion Developers
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+//! Digital signature creation and verification.
+//!
+//! # Use case:
+//! `orion::signer` can be used to cryptographically sign data and validate signatures over such data.
+//!
+//! # About:
+//! - Uses ML-DSA-65 in "hedged"/randomized mode.
+//! - Keys are based exclusively on a seed. To interact with semi-expanded/serialized signing keys, users will have to reach for [`orion::hazardous::dsa`].
+//!
+//! # Parameters:s
+//! - `m`: Message to be signed or verified a signature of.
+//! - `ctx`: Context string which must be the same on signing and verification.
+//! - `sig`: Signature to be verified.
+//!
+//! # Errors:
+//! An error will be returned if:
+//! - Verification of a signature, message and context failed.
+//! - Failure to generate random bytes securely during key-generation.
+//! - `ctx` is not `<= 255` bytes.
+//!
+//! # Example:
+//! ```rust
+//! use orion::signer::SigningKeyPair;
+//!
+//! // Randomly generate a fresh keypair
+//! let kp = SigningKeyPair::generate()?;
+//!
+//! // Sign a message
+//! let sig = kp.sign(b"Message to be signed", b"Context")?;
+//!
+//! // Verify the signature of the message and accompanying context
+//! assert!(kp.verify(b"Message to be signed", b"Context", &sig).is_ok());
+//! # Ok::<(), orion::errors::UnknownCryptoError>(())
+//! ```
+//! [`orion::signer`]: crate::signer
+//! [`orion::hazardous::dsa`]: crate::hazardous::dsa
+
+#![cfg_attr(docsrs, doc(cfg(feature = "safe_api")))]
+
+use crate::{errors::UnknownCryptoError, hazardous::dsa::mldsa65};
+
+#[derive(Debug, PartialEq)]
+/// ML-DSA-65 seed-based signing key pair.
+pub struct SigningKeyPair {
+    // We hold `SigningKey` to force the use of Seed in the higher-level API.
+    // `sk` should never be reachable by a user for this type. The ability
+    // to uses semi-expanded keys is forced down to `hazardous`.
+    sk: mldsa65::SigningKey,
+    vk: mldsa65::VerifyingKey,
+    seed: mldsa65::Seed,
+}
+
+impl SigningKeyPair {
+    /// Randomly generate a fresh ML-DSA-65 keypair.
+    pub fn generate() -> Result<Self, UnknownCryptoError> {
+        let seed = mldsa65::Seed::generate()?;
+        let mldsakp = mldsa65::KeyPair::try_from(&seed)?;
+
+        Ok(Self {
+            sk: mldsakp.signing_key,
+            vk: mldsakp.verifying_key,
+            seed,
+        })
+    }
+
+    /// Get a reference to this [`SigningKeyPair`]'s private seed.
+    pub fn private(&self) -> &mldsa65::Seed {
+        &self.seed
+    }
+
+    /// Get a reference to this [`SigningKeyPair`]'s public verifying key.
+    pub fn public(&self) -> &mldsa65::VerifyingKey {
+        &self.vk
+    }
+
+    /// Sign a message, with optional context (can be empty), using ML-DSA-65 hedged/randomized signing.
+    pub fn sign(&self, m: &[u8], ctx: &[u8]) -> Result<mldsa65::Signature, UnknownCryptoError> {
+        self.sk.sign(m, ctx)
+    }
+
+    /// Verify a signature over a message, with optional context (can be empty), using ML-DSA-65.
+    /// Returns nothing on success and [`UnknownCryptoError`] of verification failed.
+    pub fn verify(
+        &self,
+        m: &[u8],
+        ctx: &[u8],
+        sig: &mldsa65::Signature,
+    ) -> Result<(), UnknownCryptoError> {
+        self.vk.verify(m, ctx, sig)
+    }
+}
+
+impl TryFrom<&mldsa65::Seed> for SigningKeyPair {
+    type Error = UnknownCryptoError;
+
+    fn try_from(value: &mldsa65::Seed) -> Result<Self, Self::Error> {
+        let mldsakp = mldsa65::KeyPair::try_from(value)?;
+
+        Ok(Self {
+            sk: mldsakp.signing_key,
+            vk: mldsakp.verifying_key,
+            seed: mldsa65::Seed::from(value.data.bytes),
+        })
+    }
+}